==> src/24 alpha_factor_library/00_indicator_zoo.py <==
# The Indicator Zoo
#
# Source:
#
# Green, Jeremiah and Hand, John R. M. and Zhang, Frank, [The Characteristics that Provide Independent Information
# about Average U.S. Monthly Stock Returns](https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2262374), The Review
# of Financial Studies, Volume 30, Issue 12, December 2017, Pages 4389–4436

# In his 2011 American Finance Association Presidential address, John H. Cochrane (2011) challenged researchers to
# identify the firm characteristics that provide independent information about average U.S. stock returns.
#
# Cochrane issued his challenge because of the “veritable zoo” of hundreds of characteristics that have been presented
# as statistically significant predictors of the cross-section of returns to the anomaly's literature since 1970.

# | Acronym    | Author(s)                           | Date, Journal | Definition of the characteristic-based anomaly variable                                                                                                                                                                                                                                 |
# |------------|-------------------------------------|---------------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
# | acc        | Sloan                               | 1996, TAR     | Annual income before extraordinary items (ib) minus operating cash flows (oancf) divided by average total assets (at); if oancf is missing then set to change in act - change in che - change in lct++ change in dlc++ change in txp-dp                                                 |
# | aeavol     | Lerman, Livnat, and Mendenhall      | 2008, WP      | Average daily trading volume (vol) for 3 days around earnings announcement minus average daily volume for 1-month ending 2 weeks before earnings announcement divided by 1-month average daily volume. Earnings announcement day from Compustat quarterly (rdq)                         |
# | age        | Jiang, Lee, and Zhang               | 2005, RAS     | Number of years since first Compustat coverage                                                                                                                                                                                                                                          |
# | agr        | Cooper, Gulen, and Schill           | 2008, JF      | Annual percent change in total assets (at)                                                                                                                                                                                                                                              |
# | baspread   | Amihud and Mendelson                | 1989, JF      | Monthly average of daily bid-ask spread divided by average of daily spread                                                                                                                                                                                                              |
# | beta       | Fama and MacBeth                    | 1973, JPE     | Estimated market beta from weekly returns and equal weighted market returns for 3 years ending month tt-1 with at least 52 weeks of returns                                                                                                                                             |
# | bm         | Rosenberg, Reid, and Lanstein       | 1985, JPM     | Book value of equity (ceq) divided by end of fiscal year-end market capitalization                                                                                                                                                                                                      |
# | bm_ia      | Asness, Porter, and Stevens         | 2000, WP      | Industry adjusted book-to-market ratio                                                                                                                                                                                                                                                  |
# | cash       | Palazzo                             | 2012, JFE     | Cash and cash equivalents divided by average total assets                                                                                                                                                                                                                               |
# | cashdebt   | Ou and Penman                       | 1989, JAE     | Earnings before depreciation and extraordinary items (ib++dp) divided by avg. total liabilities (lt)                                                                                                                                                                                    |
# | cashpr     | Chandrashekar and Rao               | 2009, WP      | Fiscal year-end market capitalization plus long-term debt (dltt) minus total assets (at) divided by cash and equivalents (che)                                                                                                                                                          |
# | cfp        | Desai, Rajgopal, and Venkatachalam  | 2004, TAR     | Operating cash flows divided by fiscal-year-end market capitalization                                                                                                                                                                                                                   |
# | cfp_ia     | Asness, Porter and Stevens          | 2000, WP      | Industry adjusted cfp                                                                                                                                                                                                                                                                   |
# | chatoia    | Soliman                             | 2008, TAR     | 2-digit SIC - fiscal-year mean-adjusted change in sales (sale) divided by average total assets (at)                                                                                                                                                                                     |
# | chcsho     | Pontiff and Woodgate                | 2008, JF      | Annual percent change in shares outstanding (csho)                                                                                                                                                                                                                                      |
# | chempia    | Asness, Porter, and Stevens         | 1994, WP      | Industry-adjusted change in number of employees                                                                                                                                                                                                                                         |
# | chfeps     | Hawkins, Chamberlin, and Daniel     | 1984, FAJ     | Mean analyst forecast in month prior to fiscal period end date from I/B/E/S summary file minus same mean forecast for prior fiscal period using annual earnings forecasts                                                                                                               |
# | chinv      | Thomas and Zhang                    | 2002, RAS     | Change in inventory (inv) scaled by average total assets (at)                                                                                                                                                                                                                           |
# | chmom      | Gettleman and Marks                 | 2006, WP      | Cumulative returns from months tt-6 to tt-1 minus months tt-12 to tt-7                                                                                                                                                                                                                  |
# | chpmia     | Soliman                             | 2008, TAR     | 2-digit SIC - fiscal-year mean adjusted change in income before extraordinary items (ib) divided by sales (sale)                                                                                                                                                                        |
# | chtx       | Thomas and Zhang                    | 2011, JAR     | Percent change in total taxes (txtq) from quartertt-4 to tt                                                                                                                                                                                                                             |
# | cinvest    | Titman, Wei, and Xie                | 2004, JFQA    | Change over one quarter in net PP&E (ppentq) divided by sales (saleq) - average of this variable for prior 3 quarters; if saleq== 0, then scale by 0.01                                                                                                                                 |
# | convind    | Valta                               | 2016, JFQA    | An indicator equal to 1 if company has convertible debt obligations                                                                                                                                                                                                                     |
# | currat     | Ou and Penman                       | 1989, JAE     | Current assets / current liabilities                                                                                                                                                                                                                                                    |
# | depr       | Holthausen and Larcker              | 1992, JAE     | Depreciation divided by PP&E                                                                                                                                                                                                                                                            |
# | disp       | Diether, Malloy, and Scherbina      | 2002, JF      | Standard deviation of analyst forecasts in month prior to fiscal period end date divided by the absolute value of the mean forecast; if meanest== 0, then scalar set to 1. Forecast data from I/B/E/S summary files                                                                     |
# | divi       | Michaely, Thaler, and Womack        | 1995, JF      | An indicator variable equal to 1 if company pays dividends but did not in prior year                                                                                                                                                                                                    |
# | dolvol     | Chordia, Subrahmanyam, and Anshuman | 2001, JFE     | Natural log of trading volume times price per share from month tt-2                                                                                                                                                                                                                     |
# | dy         | Litzenberger and Ramaswamy          | 1982, JF      | Total dividends (dvt) divided by market capitalization at fiscal year-end                                                                                                                                                                                                               |
# | ear        | Kishore et al.                      | 2008, WP      | Sum of daily returns in three days around earnings announcement. Earnings announcement from Compustat quarterly file (rdq)                                                                                                                                                              |
# | egr        | Richardson et al.                   | 2005, JAE     | Annual percent change in book value of equity (ceq)                                                                                                                                                                                                                                     |
# | ep         | Basu                                | 1977, JF      | Annual income before extraordinary items (ib) divided by end of fiscal year market cap                                                                                                                                                                                                  |
# | fgr5yr     | Bauman and Dowen                    | 1988, FAJ     | Most recently available analyst forecasted 5-year growth                                                                                                                                                                                                                                |
# | gma        | Novy-Marx                           | 2013, JFE     | Revenues (revt) minus cost of goods sold (cogs) divided by lagged total assets (at)                                                                                                                                                                                                     |
# | grCAPX     | Anderson and Garcia-Feijoo          | 2006, JF      | Percent change in capital expenditures from yeartt-2 to year tt                                                                                                                                                                                                                         |
# | grltnoa    | Fairfield, Whisenant, and Yohn      | 2003, TAR     | Growth in long-term net operating assets                                                                                                                                                                                                                                                |
# | herf       | Hou and Robinson                    | 2006, JF      | 2-digit SIC - fiscal-year sales concentration (sum of squared percent of sales in industry for each company).                                                                                                                                                                           |
# | hire       | Bazdresch, Belo, and Lin            | 2014, JPE     | Percent change in number of employees (emp)                                                                                                                                                                                                                                             |
# | idiovol    | Ali, Hwang, and Trombley            | 2003, JFE     | Standard deviation of residuals of weekly returns on weekly equal weighted market returns for 3 years prior to month end                                                                                                                                                                |
# | ill        | Amihud                              | 2002, JFM     | Average of daily (absolute return / dollar volume).                                                                                                                                                                                                                                     |
# | indmom     | Moskowitz and Grinblatt             | 1999, JF      | Equal weighted average industry 12-month returns                                                                                                                                                                                                                                        |
# | invest     | Chen and Zhang                      | 2010, JF      | Annual change in gross property, plant, and equipment (ppegt) ++ annual change in inventories (invt) all scaled by lagged total assets (at)                                                                                                                                             |
# | IPO        | Loughran and Ritter                 | 1995, JF      | An indicator variable equal to 1 if first year available on CRSP monthly stock file                                                                                                                                                                                                     |
# | lev        | Bhandari                            | 1988, JF      | Total liabilities (lt) divided by fiscal year-end market capitalization                                                                                                                                                                                                                 |
# | lgr        | Richardson et al.                   | 2005, JAE     | Annual percent change in total liabilities (lt)                                                                                                                                                                                                                                         |
# | maxret     | Bali, Cakici, and Whitelaw          | 2011, JFE     | Maximum daily return from returns during calendar monthtt-1                                                                                                                                                                                                                             |
# | mom12m     | Jegadeesh                           | 1990, JF      | 11-month cumulative returns ending one month before month end                                                                                                                                                                                                                           |
# | mom1m      | Jegadeesh and Titman                | 1993, JF      | 1-month cumulative return                                                                                                                                                                                                                                                               |
# | mom36m     | Jegadeesh and Titman                | 1993, JF      | Cumulative returns from monthstt-36 to tt-13                                                                                                                                                                                                                                            |
# | ms         | Mohanram                            | 2005, RAS     | Sum of 8 indicator variables for fundamental performance                                                                                                                                                                                                                                |
# | mve        | Banz                                | 1981, JFE     | Natural log of market capitalization at end of month tt-1                                                                                                                                                                                                                               |
# | mve_ia     | Asness, Porter, and Stevens         | 2000, WP      | 2-digit SIC industry-adjusted fiscal year-end market capitalization                                                                                                                                                                                                                     |
# | nanalyst   | Elgers, Lo, and Pfeiffer            | 2001, TAR     | Number of analyst forecasts from most recently available I/B/E/S summary files in month prior to month of portfolio formation. nanalyst set to zero if not covered in I/B/E/S summary file                                                                                              |
# | nincr      | Barth, Elliott, and Finn            | 1999, JAR     | Number of consecutive quarters (up to eight quarters) with an increase in earnings (ibq) over same quarter in the prior year                                                                                                                                                            |
# | operprof   | Fama and French                     | 2015, JFE     | Revenue minus cost of goods sold - SG&A expense - interest expense divided by lagged common shareholders’ equity                                                                                                                                                                        |
# | orgcap     | Eisfeldt and Papanikolaou           | 2013, JF      | Capitalized SG&A expenses                                                                                                                                                                                                                                                               |
# | pchcapx_ia | Abarbanell and Bushee               | 1998, TAR     | 2-digit SIC - fiscal-year mean-adjusted percent change in capital expenditures (capx)                                                                                                                                                                                                   |
# | pctacc     | Hafzalla, Lundholm, and Van Winkle  | 2011, TAR     | Same as acc except that the numerator is divided by the absolute value of ib; if ib== 0 then ib set to 0.01 for denominator                                                                                                                                                             |
# | pricedelay | Hou & Moskowitz                     | 2005, RFS     | The proportion of variation in weekly returns for 36 months ending in monthtt explained by 4 lags of weekly market returns incremental to contemporaneous market return                                                                                                                 |
# | quick      | Ou and Penman                       | 1989, JAE     | (current assets - inventory) / current liabilities                                                                                                                                                                                                                                      |
# | rd         | Eberhart, Maxwell, and Siddique     | 2004, JF      | An indicator variable equal to 1 if R&D expense as a percentage of total assets has an increase greater than 5%.                                                                                                                                                                        |
# | rd_mve     | Guo, Lev, and Shi                   | 2006, JBFA    | R&D expense divided by end-of-fiscal-year market capitalization                                                                                                                                                                                                                         |
# | realestate | Tuzel                               | 2010, RFS     | Buildings and capitalized leases divided by gross PP&E                                                                                                                                                                                                                                  |
# | retvol     | Ang et al.                          | 2006, JF      | Standard deviation of daily returns from month tt-1                                                                                                                                                                                                                                     |
# | roaq       | Balakrishnan, Bartov, and Faurel    | 2010, JAE     | Income before extraordinary items (ibq) divided by one quarter lagged total assets (atq)                                                                                                                                                                                                |
# | roavol     | Francis et al.                      | 2004, TAR     | Standard deviation for 16 quarters of income before extraordinary items (ibq) divided by average total assets (atq)                                                                                                                                                                     |
# | roeq       | Hou, Xue, and Zhang                 | 2015 RFS      | Earnings before extraordinary items divided by lagged common shareholders’ equity                                                                                                                                                                                                       |
# | roic       | Brown and Rowe                      | 2007, WP      | Annual earnings before interest and taxes (ebit) minus nonoperating income (nopi) divided by non-cash enterprise value (ceq++lt-che)                                                                                                                                                    |
# | rsup       | Kama                                | 2009, JBFA    | Sales from quarter t minus sales from quarter tt-4 (saleq) divided by fiscal-quarter-end market capitalization (cshoq * prccq)                                                                                                                                                          |
# | salecash   | Ou and Penman                       | 1989, JAE     | Annual sales divided by cash and cash equivalents                                                                                                                                                                                                                                       |
# | secured    | Valta                               | 2016, JFQA    | Total liability scaled secured debt                                                                                                                                                                                                                                                     |
# | sfe        | Elgers, Lo, and Pfeiffer            | 2001, TAR     | Analysts mean annual earnings forecast for nearest upcoming fiscal year from most recent month available prior to month of portfolio formation from I/B/E/S summary files scaled by price per share at fiscal quarter end                                                               |
# | sgr        | Lakonishok, Shleifer, and Vishny    | 1994, JF      | Annual percent change in sales (sale)                                                                                                                                                                                                                                                   |
# | sin        | Hong & Kacperczyk                   | 2009, JFE     | An indicator variable equal to 1 if a company’s primary industry classification is in smoke or tobacco, beer or alcohol, or gaming                                                                                                                                                      |
# | SP         | Barbee, Mukherji, and Raines        | 1996, FAJ     | Annual revenue (sale) divided by fiscal year-end market capitalization                                                                                                                                                                                                                  |
# | std_dolvol | Chordia, Subrahmanyam, and Anshuman | 2001, JFE     | Monthly standard deviation of daily dollar trading volume                                                                                                                                                                                                                               |
# | stdcf      | Huang                               | 2009, JEF     | Standard deviation for 16 quarters of cash flows divided by sales (saleq); if saleq== 0, then scale by 0.01. Cash flows defined as ibq minus quarterly accruals                                                                                                                         |
# | sue        | Rendelman, Jones, and Latane        | 1982, JFE     | Unexpected quarterly earnings divided by fiscal-quarter-end market cap. Unexpected earnings is I/B/E/S actual earnings minus median forecasted earnings if available, else it is the seasonally differenced quarterly earnings before extraordinary items from Compustat quarterly file |
# | tang       | Almeida and Campello                | 2007, RFS     | Cash holdings ++ 0.715 ×× receivables ++0.547 ×× inventory ++ 0.535 ×× PPE/ totl assets                                                                                                                                                                                                 |
# | tb         | Lev and Nissim                      | 2004, TAR     | Tax income, calculated from current tax expense divided by maximum federal tax rate, divided by income before extraordinary items                                                                                                                                                       |
# | turn       | Datar, Naik, and Radcliffe          | 1998, JFM     | Average monthly trading volume for most recent 3 months scaled by number of shares outstanding in current month                                                                                                                                                                         |
# | zerotrade  | Liu                                 | 2006, JFE     | Turnover weighted number of zero trading days for most recent 1 month                                                                                                                                                                                                                   |
